skills/video-generator-skill/requirements.txt
# Skill Template Dependencies
# 核心依赖
requests>=2.28.0

# 可选依赖 - 美化输出
rich>=13.0.0
tqdm>=4.65.0

# 可选依赖 - 环境变量管理
python-dotenv>=1.0.0
